WARNING: This merge brings in upstream changes to the Clang
command-line options and environment variables needed for offloading
support.  Specifically, instead of specifying the location of
`libmoptarget-nvptx-*.bc` files via `LIBRARY_PATH`, it should be
specified via `--libomptarget-nvptx-bc-path=`.  In our experience,
this usually affects Clacc when working from a build directory rather
than an install directory, and this merge updates the section "Usage
from a Build Directory" in Clacc's `README.md` accordingly.

This merge brings in the following conflicting commits from upstream:

* 2a581710c194: This commit makes the above changes for locating
  `libmoptarget-nvptx-*.bc`.  Besides updating Clacc's documentation,
  this merge updates the lit configuration for Clacc's libomptarget
  `multitarget` test suite, Clang OpenACC test suite, and libacc2omp
  test suite, mimicking similar changes made by this commit for
  upstream's libomptarget `nvptx64-nvidia-cuda` test suite.
* c7d7b98e5263, 7efb825382c0, 347f3c186d3f, 31cca215654a,
  000875c127c7: These commits define a new AST `Stmt` node and thus
  insert new enumerators into `CXCursorKind` immediately before
  `CXCursor_LastStmt`.  Clacc does the same.  This merges combines
  those, keeping Clacc's enumerators last.  Like the previous time
  this happened, this change causes `CXCursor_LastStmt` to collide
  with the existing `CXCursor_TranslationUnit` value, which this merge
  thus increases in both `clang/include/clang-c/Index.h` and
  `clang/bindings/python/clang/cindex.py`.  To avoid conflicts the
  next this happens, it increases it from 301 to 350.
* c3aecf87d5b9: This commit changes libomptarget's `Devices` vector to
  store `std::unique_ptr<DeviceTy>` instead of `DeviceTy`, and it
  removes the `DeviceTy` copy constructor and copy assignment operator
  because they are then unecessary.  Clacc adds various uses of that
  vector, and this merge updates those uses for the new type.  Clacc
  had also extended the copy constructor and copy assignment operator
  for `DeviceTy` members that Clacc adds, and this merge resolves the
  resulting conflict in favor of upstream: just remove the copy
  constructor and copy assignment operator as no longer needed.
* This merge also resolves contextual conflicts with other commits.

This merge seems to have a fixed one upstream Clang bug brought in by
merge commit 97a21f56799c and described in the commit log there: fatal
address space errors in the case of -O1.  Thus, this merge backs out
the workaround in Clacc's test suite described there: -O2.  Sadly,
instead of the compile error, we're now occasionally seeing run-time
CUDA illegal memory access errors from Clacc's test suite.  So far,
they seem to be more common at -O2 than -O1, further motivating
reversing the previous workaround in the test suite.

+++ openmp/libomptarget/src/rtl.h
@@ -64,7 +64,7 @@ struct RTLInfoTy {
       uint64_t, __tgt_async_info *
       OMPT_SUPPORT_IF(, const ompt_plugin_api_t *ompt_api));
   typedef int64_t(init_requires_ty)(int64_t);
-  typedef int64_t(synchronize_ty)(int32_t, __tgt_async_info *);
+  typedef int32_t(synchronize_ty)(int32_t, __tgt_async_info *);
   typedef int32_t (*register_lib_ty)(__tgt_bin_desc *);
   typedef int32_t(supports_empty_images_ty)();
   typedef void(print_device_info_ty)(int32_t);
